# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sayenly API Documentation",
    "version": "1.0.0",
    "description": "Documentation for Sayenly Backend API",
    "contact": {
      "name": "Sayenly Support",
      "email": "support@sayenly.com"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "UUID": {
        "type": "string",
        "format": "uuid",
        "description": "A Universally Unique Identifier (UUID) v4",
        "example": "123e4567-e89b-12d3-a456-426614174000"
      },
...
      "DateTime": {
        "type": "string",
        "format": "date-time",
        "description": "A date-time string in ISO 8601 format",
        "example": "2023-01-23T14:30:00Z"
      },
      "URL": {
        "type": "string",
        "format": "uri",
        "description": "A valid URL",
        "example": "https://example.com/resource"
      },
      "JSON": {
        "type": "object",
        "description": "A JSON object",
        "additionalProperties": true,
        "example": {
          "key": "value",
          "nested": {
            "data": 123
          }
        }
      },
...
      "Project": {
        "type": "object",
        "required": [
          "id",
          "headline",
          "description",
          "images",
          "address",
          "date",
          "htmlContent",
          "inApp",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/UUID"
          },
          "headline": {
            "type": "string",
            "description": "Main headline for the project"
          },
          "description": {
            "type": "string",
            "description": "Description of the project"
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/URL"
            },
            "description": "List of image URLs for the project"
          },
          "address": {
            "type": "string",
            "description": "Location address of the project"
          },
          "date": {
            "$ref": "#/components/schemas/DateTime"
          },
          "inApp": {
            "type": "boolean",
            "description": "Whether the project is visible in the app"
          },
          "htmlContent": {
            "$ref": "#/components/schemas/JSON",
            "description": "Rich text content for the project"
          },
          "services": {
            "type": "array",
            "description": "Services related to this project",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "$ref": "#/components/schemas/UUID"
                },
                "name": {
                  "type": "string"
                }
              }
            }
          },
          "areas": {
            "type": "array",
            "description": "Areas related to this project",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "$ref": "#/components/schemas/UUID"
                },
                "name": {
                  "type": "string"
                }
              }
            }
          },
          "createdAt": {
            "$ref": "#/components/schemas/DateTime"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/DateTime"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/project/get-all": {
      "get": {
        "summary": "Get all projects",
        "tags": [
          "Project"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of all projects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/project/get-by-id": {
      "get": {
        "summary": "Get a project by ID",
        "tags": [
          "Project"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UUID"
            },
            "description": "The project ID"
          }
        ],
        "responses": {
          "200": {
            "description": "The project information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/project/create": {
      "post": {
        "summary": "Create a new project",
        "tags": [
          "Project"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "headline",
                  "description",
                  "images",
                  "address",
                  "date",
                  "htmlContent"
                ],
                "properties": {
                  "headline": {
                    "type": "string",
                    "description": "Main headline for the project"
                  },
                  "description": {
                    "type": "string",
                    "description": "Description of the project"
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/URL"
                    },
                    "description": "List of image URLs for the project"
                  },
                  "address": {
                    "type": "string",
                    "description": "Location address of the project"
                  },
                  "date": {
                    "$ref": "#/components/schemas/DateTime"
                  },
                  "inApp": {
                    "type": "boolean",
                    "default": false,
                    "description": "Whether the project is visible in the app"
                  },
                  "htmlContent": {
                    "$ref": "#/components/schemas/JSON",
                    "description": "Rich text content for the project"
                  },
                  "services": {
                    "type": "array",
                    "description": "Services to associate with this project",
                    "items": {
                      "type": "object",
                      "required": [
                        "id"
                      ],
                      "properties": {
                        "id": {
                          "$ref": "#/components/schemas/UUID"
                        }
                      }
                    }
                  },
                  "areas": {
                    "type": "array",
                    "description": "Areas to associate with this project",
                    "items": {
                      "type": "object",
                      "required": [
                        "id"
                      ],
                      "properties": {
                        "id": {
                          "$ref": "#/components/schemas/UUID"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The created project",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/project/update": {
      "put": {
        "summary": "Update a project",
        "tags": [
          "Project"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UUID"
            },
            "description": "The project ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "headline": {
                    "type": "string",
                    "description": "Main headline for the project"
                  },
                  "description": {
                    "type": "string",
                    "description": "Description of the project"
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/URL"
                    },
                    "description": "List of image URLs for the project"
                  },
                  "address": {
                    "type": "string",
                    "description": "Location address of the project"
                  },
                  "date": {
                    "$ref": "#/components/schemas/DateTime"
                  },
                  "inApp": {
                    "type": "boolean",
                    "description": "Whether the project is visible in the app"
                  },
                  "htmlContent": {
                    "$ref": "#/components/schemas/JSON",
                    "description": "Rich text content for the project"
                  },
                  "services": {
                    "type": "array",
                    "description": "Services to associate with this project",
                    "items": {
                      "type": "object",
                      "required": [
                        "id"
                      ],
                      "properties": {
                        "id": {
                          "$ref": "#/components/schemas/UUID"
                        }
                      }
                    }
                  },
                  "areas": {
                    "type": "array",
                    "description": "Areas to associate with this project",
                    "items": {
                      "type": "object",
                      "required": [
                        "id"
                      ],
                      "properties": {
                        "id": {
                          "$ref": "#/components/schemas/UUID"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated project",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/project/delete": {
      "delete": {
        "summary": "Delete a project",
        "tags": [
          "Project"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UUID"
            },
            "description": "The project ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Project deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  }
}